\name{aljazeera}
\alias{aljazeera}
\docType{data}
\title{
Arabic text
}
\description{
Arabic text from the front page of Aljazeera
}
\usage{data("aljazeera")}
\format{
  Arabic text data
}
\source{
http://aljazeera.net/portal
}
\examples{
## Load data

data(aljazeera)
}
\keyword{datasets}
